move TasksList to Widgets main directory
